added text atoms
when a text is stored it gets split up into atoms (words or nonwords).
when the study page is calles, the text will be build with these atoms
instead of the whole text. with this change, i can now add information
(annotation, translation, case) to a specific word of a text in the
database.